Create CAP-Zone-Recovery-EA-Pro-MT5
This code represents the basic structure and concept of the Zone Recovery algorithm. It is important to note that before using this kind of robots in real trading, they should be thoroughly tested on historical data and in demo account mode to make sure that they are effective and safe.